Prior to this commit, local dates (i.e. 2011-01-01) were encoded in the traditional java style as the time at midnight on the given day. The problem with this is that that time depends on the server/browser's timezone and so when transmitting Dates encoded as posix times (milliseconds since midnight 1970-01-01 UTC) between a browser on UTC+2 and the server on UTC, 2011-jan-15 00:00 became 2011-jan-14 22:00, which is not right!
This problem was magnified because the dates were then stored in locally in sqlite as posix times, which
provided more opportuntites for dates to be rolled back.

From this commit on, we distingush betwen java.util.Date, which is used to indicate a point in time, such
as the moment at which a record was edited or the last login date of a user, and com.bedatadriven.rebar.time.calendar.LocalDate,
which is modelled after the new javax.time API (JSR310) and indicates a day in the ISO-8601 date system with
no reference to time zone. (See http://sourceforge.net/apps/mediawiki/threeten/index.php?title=ThreeTen for more info)