Fixes bug for Admin and PartnerFilterPanel that caused valid choices to be hidden
The FilterPanels, like those in Excel, only show choices given the other filter choices.
So if you choose to filter by AVSI, it should only give you a choice to filter by the provinces
in which avsi is active and not all of them.

However, when reports were loaded, the filter associated with the report contains restrictions
for all dimensions, and this was applied as part of the base filter, so you would only see those
provinces/partners that had already been selected.

This commit ensures that the two affected filter Panels remove their own dimension from the base filter
before applying.