Fixes (again) the issue with ajax-initiated downloads
The workaround attempted in 3b13c7421c92c1758950f9437a17fe555ca58911
didn't work at all outside of FireFox, so this commits adds a browser-
differentiated approach.

For FireFox 3.5+ and Chrome, downloads are initiated smoothly with iframe.
For all other browsers, user is prsented with popup box and a link that
is sure to work.