Properly fixes ajax download issues.
For exports, adds a dialog which provides feedback to user during server-side generation phase,
always tries to initiate donwload with iframe, but always offers link as fallback.

For attachments, add GET method to Servlet that allows the UI to send the user to an endpoint
directly following a click. objct is proxied through app server.