Refactored UserEditor and corrected password validation regexp
This commit removes an early pattern involving MVP and several layers of
inheritance, at least for the UserDbEditor.

While MVP promises more testability and cleaner code, it just doesn't work
very well with GXT and adding inheritance resulted in trainwrecks of code.

It would be nice to ultimately refactor all of the remaining Crud pages
that use the ugly AbstractGridPresenter/AbstractGridPanel pattern but as
long as the code is running it's probably not a priority.