Improves handling of incompatible remote serialization exceptions.
If we are unable to communicate with the server because of class changes
in GWT-RPC, then we want to prompt the user to refresh the page immediately
in order to get the latest version of the client.

However, if the AppCache is active, refreshing the page will NOT load the
new version and we will be back to square one. This commit adds a UI that will
forcibly clear the AppCache before refreshing the page if necessary so that
the new version can be loaded immediately.