Removed unused fields dateEdited and dateCreated from the Location table
Location objects are now append-only as they are shared between users,
so there is no need for seperate dateCreated and dateEdited fields.

Also, the timeEdited field had been added awhile back to workaround
the date-time resolution limit of MySQL so we don't need either
of the dateCreated or dateEdited fields.

Finally, this commit adds a unique constraint on timeEdited to avoid
problems with synchronization, as this field is used as a sequence
index during synchronization. We should probably move where the
sequence numbers are generated properly, but this will do for the time
being.