Adds a wizard-like dialog to create the initial map
With the rollout in Mali, there were problems because the "default" basemap is
specific to the DRC. By adding a wizard like dialog to select the intial layer
before navigating the maps page, we have a chance to select the default basemap
based on the the indicators chosen.

I think it's also nice to avoid starting with a blank map, which I think may
confuse some users.