Optimizes the GetSites query.
With a growing number of linked indicators, the query had slowed down to a crawl.
This optimizes the query by applying the joins and filters to both "real" sites
and "linked" sites individually, unioning the results, and then sorting and apply limits.

We also take advantage of the SQL_CALC_FOUND_ROWS keyword which allows to get the total
record count without having to reexecute the query, at least with MySQL.