Fixes problem with appcache affecting users with multiple sign ins
The SelectionScriptServlet had been returning a redirect script for ActivityInfo.nocache.js in
the event that the request was not authenticated.

However, this script was getting cached by the browser and getting stuck. Even after the
user authenticated, the manifest does not change, so the browser does not try to update
the cache and the redirect script gets stuck.