Fixes nasty StackOverflowError that was crashing our beanstalk instances
The LocationUpdateBuilder used an 'in' clause in a hibernate query to fetch
additional information about the update locations. Hibernate apparently uses
recursion to process the in clause, which lead to a SOE when we encountered
the newly updated mali locations with 20k + locations pre-added.

This is apparently a known bug, which is marked as closed but clearly still
not fixed, given that we're using the latest version of hibernate (3.4.0)

https://hibernate.onjira.com/browse/HHH-2166

The big problem with the SOE is that it totally crashes the thread so the
exception is not even logged, and the open connection is not closed, which
was ended up sinking our servers when bonecp eventually hit the max
connection.