Fixes problem encountered by offline users when login is not saved
Offline mode currently relies on the authToken cookie to determine
which user's database to load. If this cookie is not set to be
persisted across sessions, then if the user tries to log back in
offline, there will be no user chosen and the user will
be redirected to the (online) login form.