Removed Quartz dependency
The version of quartz on which ActivityInfo previously depended
conflicted with the version of quartz used by the AppEngine dev
server to simulate the task queue api during development.

We had previously used the quartz library to handle nightly
emailing of subscribed reports, but since the migration to
AppEngine in 2.7, we take advantage of the AppEngine cron service.

I had left the quartz code in place with the idea that it would
"be nice" if 2.7 could also be run in a standalone java servlet,
but we currently don't have the time/resources to support this use case,
so we're just going to rip it out rather than fight with the
dev server.